demo/client/bridge/pyproject.toml
[project]
name = "op-bridge"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "mcp==1.4.1",
    "httpx==0.28.1",
    "python-dotenv==1.0.1"
]
